+++ packages/benchmark-js/test/benchmark.ts
@@ -1,6 +1,6 @@
-import { BenchmarkGroup } from '../src';
+import { Benchmark } from '../src';
 
-const benchmark = new BenchmarkGroup();
+const benchmark = new Benchmark();
 benchmark.setEmptyTest();
 benchmark.add('Benchmark-1', () => {});
 benchmark.add('Benchmark-2', (arg1, arg2, arg3) => arg3, {